requirements.txt
streamlit==1.37.0 
openai==1.35.13 
langchain==0.2.11 
langchain-community==0.2.10 
langchain-chroma==0.1.1 
chromadb==0.5.0 
sentence-transformers==3.0.1 
pypdf2==3.0.1 
python-docx==1.1.0 
faiss-cpu==1.8.0 
networkx==3.3 
tiktoken==0.7.0